Separate types and data to fix undefined types issue js_only
