+++ assets/prototype/application/services/config/index.ts
@@ -1,0 +1,5 @@
+export { default as useConfig } from './useConfig';
+
+export * from './ConfigData';
+
+export * from './types';
