+++ assets/prototype/application/services/config/useConfig.ts
@@ -1,6 +1,6 @@
 import { useContext } from 'react';
 
-import { ConfigDataProps } from './ConfigData';
+import { ConfigDataProps } from './types';
 import { ConfigContext } from '../context/ConfigProvider';
 
 export const useConfig = (): ConfigDataProps => useContext<ConfigDataProps>(ConfigContext);
